+++ lib/oauth/client.ts
@@ -1,22 +1,4 @@
 /**
- * (c) Copyright Ascensio System SIA 2025
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @license
- */
-
-/**
  * @module
  * @mergeModuleWith oauth
  */
@@ -26,6 +8,7 @@
 
 import contentType from "content-type"
 import * as z from "zod"
+import * as errors from "../util/errors.ts"
 import * as r from "../util/result.ts"
 import type {
 	AuthorizeRequest,
@@ -315,21 +298,40 @@ export async function checkResponse(req: Request, res: Response): Promise<r.Resu
 	await (async() => {
 		let h = res.headers.get("Content-Type")
 		if (!h) {
+			let err = new Error("Content-Type is missing")
+			o.error = "server_error"
+			o.error_description = errors.format(err)
 			return
 		}
 
 		let p = r.safeSync(contentType.parse, h)
-		if (p.err || p.v.type !== "application/json") {
+		if (p.err) {
+			let err = new Error("Parsing Content-Type", {cause: p.err})
+			o.error = "server_error"
+			o.error_description = errors.format(err)
+			return
+		}
+
+		if (p.v.type !== "application/json") {
+			let err = new Error("Content-Type is invalid")
+			o.error = "server_error"
+			o.error_description = errors.format(err)
 			return
 		}
 
 		let c = r.safeSync(res.clone.bind(res))
 		if (c.err) {
+			let err = new Error("Cloning response", {cause: c.err})
+			o.error = "server_error"
+			o.error_description = errors.format(err)
 			return
 		}
 
 		let b = await r.safeAsync(c.v.json.bind(c.v))
 		if (b.err) {
+			let err = new Error("Parsing response", {cause: b.err})
+			o.error = "server_error"
+			o.error_description = errors.format(err)
 			return
 		}