+++ test/test.ts
@@ -1,21 +1,3 @@
-/**
- * (c) Copyright Ascensio System SIA 2025
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @license
- */
-
 /* eslint-disable typescript/consistent-type-definitions */
 
 import assert from "node:assert/strict"
@@ -862,8 +844,6 @@ void test.suite("validates config", () => {
 
 void test("oauth server", (t) => {
 	// todo: client.ts
-	// upstream can respond without Content-Type
-	// upstream can respond with invalid Content-Type
 	// upstream can respond with invalid JSON
 
 	// todo: errors
@@ -2581,6 +2561,113 @@ void test("oauth server", (t) => {
 					let eb: object = {
 						error: "server_error",
 						error_description: "some_reason",
+					}
+
+					assert.deepEqual(ab.v, eb)
+				}
+
+				await Promise.race([hp, tf()])
+			})
+
+			void t.test("returns error when upstream responds without Content-Type header", async(t) => {
+				let [hs, ha] = await setupHttp(t)
+
+				let hp = onRequest(t, hs, (_, res) => {
+					res.statusCode = 418
+					res.end()
+				})
+
+				let tf = async(): Promise<void> => {
+					let e: object = {
+						DOCSPACE_OAUTH_BASE_URL: `http://[${ha.address}]:${ha.port}/`,
+					}
+
+					let a = await setup(t, e)
+
+					let u = r.safeNew(URL, o.path, `http://[${a.address}]:${a.port}/`)
+					assert.ok(u.err === undefined)
+
+					let f = new URLSearchParams()
+
+					for (let [k, v] of Object.entries(o.body)) {
+						f.set(k, v)
+					}
+
+					let i: RequestInit = {
+						method: "POST",
+						headers: {
+							"Content-Type": "application/x-www-form-urlencoded",
+						},
+						body: f.toString(),
+					}
+
+					let fetch = withAuth(globalThis.fetch)
+
+					let res = await r.safeAsync(fetch, u.v, i)
+					assert.ok(res.err === undefined)
+
+					assert.ok(res.v.status === 418)
+
+					let ab = await readFetchJson(res.v)
+					assert.ok(ab.err === undefined)
+
+					let eb: object = {
+						error: "server_error",
+						error_description: "Content-Type is missing",
+					}
+
+					assert.deepEqual(ab.v, eb)
+				}
+
+				await Promise.race([hp, tf()])
+			})
+
+			void t.test("returns error when upstream responds with invalid Content-Type header", async(t) => {
+				let [hs, ha] = await setupHttp(t)
+
+				let hp = onRequest(t, hs, (_, res) => {
+					res.statusCode = 418
+					res.setHeader("Content-Type", "text/plain")
+					res.end()
+				})
+
+				let tf = async(): Promise<void> => {
+					let e: object = {
+						DOCSPACE_OAUTH_BASE_URL: `http://[${ha.address}]:${ha.port}/`,
+					}
+
+					let a = await setup(t, e)
+
+					let u = r.safeNew(URL, o.path, `http://[${a.address}]:${a.port}/`)
+					assert.ok(u.err === undefined)
+
+					let f = new URLSearchParams()
+
+					for (let [k, v] of Object.entries(o.body)) {
+						f.set(k, v)
+					}
+
+					let i: RequestInit = {
+						method: "POST",
+						headers: {
+							"Content-Type": "application/x-www-form-urlencoded",
+						},
+						body: f.toString(),
+					}
+
+					let fetch = withAuth(globalThis.fetch)
+
+					let res = await r.safeAsync(fetch, u.v, i)
+					assert.ok(res.err === undefined)
+
+					assert.ok(res.v.status === 418)
+
+					let ab = await readFetchJson(res.v)
+					assert.ok(ab.err === undefined)
+
+					let eb: object = {
+						error: "server_error",
+						error_description: "Content-Type is invalid",
 					}
 
 					assert.deepEqual(ab.v, eb)