+++ CHANGELOG.md
@@ -17,7 +17,8 @@ This document records all notable changes to the project, following the
   `express-rate-limit` to 8.2.1, `jsonwebtoken` to 9.0.3, `zod` to 4.1.13
   ([1da5a12], [05d37dd]);
 - Update Node.js to 24.11.1 in OCI image ([02b1525]);
-- Remove `enum` from output schemas ([ed5a178]).
+- Remove `enum` from output schemas ([ed5a178]);
+- Change project license from Apache 2.0 to MIT ([b6a3923]).
 
 ### Fixed
 
@@ -44,7 +45,9 @@ This document records all notable changes to the project, following the
 - Fix returning error instead of success response from the OAuth introspect
   endpoint when the token is not yet valid or expired ([38a3a75]);
 - Fix incorrect error status returned by the OAuth introspect endpoint when the
-  token is invalid ([38a3a75]).
+  token is invalid ([38a3a75]);
+- Fix returning empty error when the upstream OAuth server responds without or
+  with invalid Content-Type header ([5c52ebe]).
 
 ## [3.1.0] - 2025-11-28
 
@@ -309,6 +312,8 @@ There are no noticeable changes in version [3.0.1].
 [0.1.0]: https://github.com/onlyoffice/docspace-mcp/releases/tag/v0.1.0/
 
 [01bc863]: https://github.com/onlyoffice/docspace-mcp/commit/01bc863b7171c9bec9dd2fd277dbd83e710e7e6c/
+[5c52ebe]: https://github.com/onlyoffice/docspace-mcp/commit/5c52ebedf15221a5e9e7c51d1ff9348a89e36b2e/
+[b6a3923]: https://github.com/onlyoffice/docspace-mcp/commit/b6a392338b4c9e92f0364dda29212a71c65be2cd/
 [38a3a75]: https://github.com/onlyoffice/docspace-mcp/commit/38a3a7587eb8bb4a4c02c136426cf1d59e108c9d/
 [a733d43]: https://github.com/onlyoffice/docspace-mcp/commit/a733d43b8881f4825399737bd3ddb172a3b51fac/
 [658ffae]: https://github.com/onlyoffice/docspace-mcp/commit/658ffae49f0f8d253346c29f9b0bebb752329f79/